Rename temprature -> temperature.
