+++ .gitignore
@@ -1,3 +1,4 @@
 node_modules/
+.idea/
 .vscode/settings.json
 yarn.lock
