+++ specs/temp.spec.js
@@ -1,2 +1,2 @@
-import Temprature from '@/temprature'
+import Temperature from '@/temperature'
 import { mount } from '@vue/test-utils'
